Move TBAP into tiled.access_policies
